// front/app/jeddah/model/mdlJeddahParams.ts
export interface MdlJeddahParamsAddlfn {
  airlfl_addfln: string;
  flnbfl_addfln: string;
  fldtef_addfln: string;
  routfl_addfln: string;
  fltype_addfln: string;
}